moved translation functions from algebras to seperate functions
