+++ ObjectAlgebraExecutionGraphs/Algebras/CSharpTranslatableGraphAlgebra.cs
@@ -17,73 +17,6 @@ namespace ObjectAlgebraExecutionGraphs.Algebras
 
         public Type TypeFromString(string typeString) => Type.GetType(typeString);
 
-        private static IEnumerable<string> CallPureDependents(ICSharpTranslatableNode node, IEnumerable<NodeConnection<ICSharpTranslatableNode>> dataConnections)
-        {
-            for (int toIndex = 0; toIndex < node.Inputs.Count; toIndex++)
-            {
-                var conns = dataConnections.Where(conn => conn.ToNode == node && conn.ToPinIndex == toIndex).ToArray();
-                if (conns.Length > 1)
-                {
-                    throw new Exception();
-                }
-
-                if (conns.Length == 1)
-                {
-                    var conn = conns[0];
-                    if (conn.FromNode.IsPure)
-                    {
-                        foreach (var call in CallPureDependents(conn.FromNode, dataConnections))
-                        {
-                            yield return call;
-                        }
-                    }
-
-                    yield return $"{conn.FromNode.PureFunctionName}();\n";
-                }
-            }
-        }
-
-        public string TranslateImperative(IEnumerable<ICSharpTranslatableNode> nodes, IEnumerable<NodeConnection<ICSharpTranslatableNode>> dataConnections, IEnumerable<NodeConnection<ICSharpTranslatableNode>> execConnections)
-        {
-            StringBuilder builder = new StringBuilder();
-
-            foreach (var node in nodes)
-            {
-                builder.Append(node.TranslateVariables());
-            }
-
-            foreach (var pureNode in nodes.Where(node => node.IsPure))
-            {
-                builder.Append($"void {pureNode.PureFunctionName}()\n");
-                builder.Append("{\n");
-                builder.Append(pureNode.TranslatePureFunctions());
-                builder.Append("}\n");
-            }
-
-            foreach (var stateNode in nodes.Where(node => !node.IsPure))
-            {
-                IEnumerable<string> pureCalls = CallPureDependents(stateNode, dataConnections);
-
-                if (!stateNode.IsPure)
-                {
-                    var outputConnectedLabels = new string[stateNode.ExecOutputCount];
-
-                    var nodeExecConnections = execConnections
-                        .Where(conn => conn.FromNode == stateNode)
-                        .ToArray();
-
-                    foreach (var execConn in nodeExecConnections)
-                    {
-                        outputConnectedLabels[execConn.FromPinIndex] = execConn.ToNode.ExecInputs[execConn.ToPinIndex];
-                    }
-
-                    builder.Append(stateNode.TranslateStates(outputConnectedLabels, string.Concat(pureCalls.Distinct())));
-                }
-            }
-
-            return builder.ToString();
-        }
-
         private abstract class BaseCSharpTranslatableNode : ICSharpTranslatableNode
         {
             public IImmutableList<(Type type, string variableName)> Inputs { get; private set; } = ImmutableList<(Type, string)>.Empty;
