+++ ObjectAlgebraExecutionGraphs/Variants/IExecutionGraphAlgebra.cs
@@ -6,6 +6,5 @@ namespace ObjectAlgebraExecutionGraphs.Variants
     public interface IExecutionGraphAlgebra<TType, TNode>
     {
         public TNode CreateConcatenateNode();
-        public string TranslateImperative(IEnumerable<TNode> nodes, IEnumerable<NodeConnection<TNode>> dataConnections, IEnumerable<NodeConnection<TNode>> execConnections);
     }
 }
